Many many fixes
- Fixed an issue with the appbar in resources on screen resize and replaced <v-content> with <v-main>
- Fixed an issue where the icon and text of the snackbar would be on different lines
- Fixed an issue with the title not dispaying from environment variable
- Introduced loading, deleting, and saving indicators
- Made evrything except images confirm save success
- Improved the file upload tool to be more reliable and show deleting status
- Fixrd how journeys are saved
- Added deletion confirmation prompts, except on images
- Added skeleton loaders for categories, journeys, and resources
- Improved how file.js is structured

+++ src/Resources.vue
@@ -1,37 +1,40 @@
 <template>
   <div>
-    <v-navigation-drawer app absolute :clipped="true" color="blue--lighten-1" v-model="drawer">
-      <v-icon large @click.stop="drawer = !drawer"> mdi-chevron-left </v-icon>
-      <v-text-field v-model="searchText" filled label="Search" clearable ></v-text-field>
+    <v-navigation-drawer app  clipped permanent color="blue--lighten-1" :value="false" disable-resize-watcher>
+      <v-text-field v-if="!loading" v-model="searchText" filled label="Search" clearable ></v-text-field>
       <v-list>
-        <v-list-item-group v-if="resources.length" v-model="resourceIndex" color="primary">
-          <v-list-item v-for="(resource, i) in filteredResourceList" :key="i">
-            <v-list-item-content>
-              <v-list-item-title v-if="resource.doc" v-html="resource.doc.name">Hello</v-list-item-title>
-            </v-list-item-content>
+        <div v-if="loading">
+          <v-skeleton-loader type="list-item"/>
+          <v-skeleton-loader type="list-item"/>
+          <v-divider/>
+          <v-skeleton-loader v-for="n in 5" :key="n" type="list-item-avatar"/>
+        </div>
+        <div v-else>
+          <v-list-item color="primary" @click="newResource">
+            <v-list-item-icon>
+              <v-icon>mdi-plus</v-icon>
+            </v-list-item-icon>New Resource
           </v-list-item>
-        </v-list-item-group>
-        <v-list-item color="primary" @click="newResource">
-          <v-list-item-icon>
-            <v-icon>mdi-plus</v-icon>
-          </v-list-item-icon>New Resource
-        </v-list-item>
+          <v-divider/>
+          <v-list-item-group v-if="resources.length" v-model="resourceIndex" color="primary">
+            <v-list-item v-for="(resource, i) in filteredResourceList" :key="i">
+              <v-list-item-content>
+                <v-list-item-title v-if="resource.doc" v-html="resource.doc.name">Hello</v-list-item-title>
+              </v-list-item-content>
+            </v-list-item>
+          </v-list-item-group>
+        </div>
       </v-list>
     </v-navigation-drawer>
-    <v-icon large @click.stop="drawer = !drawer"> mdi-chevron-right </v-icon>
-    <v-content>
+    <v-main style="padding-left:256px">
       <v-container fluid class="fill-height" v-if="currentResource">
-        <v-container>
-          <v-btn-toggle>
-            <v-btn @click="validate">Validate</v-btn>
-            <v-btn v-if="currentResource.id" @click="updateResource" :loading="updateLoading">Update</v-btn>
-            <v-btn v-else @click="createResource">Save</v-btn>
-            <v-btn @click="confirmDelete">Delete</v-btn>
-          </v-btn-toggle>
-        </v-container>
         <component ref="resourceComponent" :is="component" v-model="currentResource.doc"/>
+        <v-container>
+            <v-btn class="mr-5" :disabled="deleting" :loading="saving" @click="saveResource">Save</v-btn>
+            <v-btn :disabled="saving" :loading="deleting" @click="confirmDelete">Delete resource</v-btn>
+        </v-container>
       </v-container>
-    </v-content>
+    </v-main>
   </div>
 </template>
 
@@ -39,6 +42,8 @@
 <script>
   import ResourceEditor from './components/controls/ResourceEditor'
   import {editorEndpoint} from '@/utils/endpoints.js'
+  import {v4 as uuidv4} from 'uuid'
+  import {savePopup} from '@/utils/ui'
   
   export default {
     components :{
@@ -49,14 +54,16 @@
       resourceIndex: -1,
       component: "resource-editor",
       endpoint: process.env.VUE_APP_API_ENDPOINT,
-      drawer: true,
       searchText: "",
-      updateLoading: false
+      saving: false,
+      deleting: false,
+      loading: true
     }),
     created() {
       fetch(editorEndpoint + '/resources')
       .then(res => res.json())
       .then(res => { this.resources = res })
+      .finally(() => {this.loading = false})
     },
     computed: {
       currentResource() {
@@ -71,71 +78,57 @@
       }
     },
     methods: {
-      validate() {
-        return this.$refs.resourceComponent.validate()
-      },
-      updateResource() {
-        if (!this.validate()) return;
-        this.updateLoading = true;
+      saveResource() {
+        this.saving = true;
+        this.$set(this.currentResource, "id", this.currentResource.id || uuidv4())
         let putReq = {
           method: 'PUT',
-          body:JSON.stringify({
-            updates:[
-              {
-                paramName: "doc", paramValue: this.currentResource.doc
-              }
-            ]
-          })
+          body:JSON.stringify(this.currentResource.doc)
         }
         fetch(`${editorEndpoint}/resources/${this.currentResource.id}`, putReq )
-          .then(() => {
-            this.$store.dispatch('doSnackbar', {text: "Changes saved successfully", colour: "success", icon: 'mdi-check-circle'})
+          .then((res) => {
+            savePopup(res.status)
           })
           .catch((err) => {
             console.error(err)
-            this.$store.dispatch('doSnackbar', {text: "Changes have not been saved", colour: "error", icon: 'mdi-alert-circle'})
+            savePopup(false)
           })
           .finally(() => {
-            this.updateLoading = false;
+            this.saving = false;
           })
       },
-      createResource() {
-        if (!this.validate()) return;
-        let req = {
-          method: "POST",
-          headers: {
-            'Content-Type': 'application/json'
-          },
-          body: JSON.stringify(this.currentResource.doc)
-        }
-        fetch(`${editorEndpoint}/resources`, req)
-        .then((res)=> res.json())
-        .then((res)=>{
-          this.$set(this.resources, this.resourceIndex, res)
-        })
-        .catch(console.log)
-      },
       confirmDelete() {
-            this.$dialog
-                .display(
-                    "Delete Resource",
-                    "Are you sure you want to delete this resource? This action cannot be undone",
-                    [{text:'Cancel', color:''}, {text:'Yes, Delete', color:''}]
-                )
-                .then((result) => {
-                    if (result === 1) this.deleteResource();
-                });
+        this.$dialog
+            .display(
+                "Delete Resource",
+                "Are you sure you want to delete this resource? This action cannot be undone",
+                [{text:'No', color:''}, {text:'Yes, Delete', color:''}]
+            )
+            .then((result) => {
+                if (result === 1) {
+                  this.deleting = true;
+                  this.deleteResource();
+                }
+            });
         },
       deleteResource() {
+        if (!this.currentResource.id) {
+          this.removeResUi()
+          return
+        }
+
+        this.deleting = true
         let delReq = {
           method: "DELETE"
         }
         fetch(`${editorEndpoint}/resources/${this.currentResource.id}`, delReq)
-        .then(()=>{
+        .then(() => this.removeResUi)
+        .finally(() => {this.deleting = false})
+      },
+      removeResUi() {
           this.resources.splice(this.resourceIndex,1);
           this.resourceIndex = null;
           this.deleteConfirmation = false
-        })
       },
       newResource() {
         let item = {
